corpus: the banked test sources catch the r48c emission -- proven by the sweep that doubles as the L3 order oracle
The full 110-package validated sweep against the L3 corpus: 110/110, 13,628
verdicts, 0 FAIL -- which is both the ruled empirical check that the layout
change moved nothing order-sensitive (os and every other reordered package's
suite included) and the proof behind this bank: 50 test sources pick up the
warning-roots emission (dead named-result prologues gone, the CS8778 tables
in their Go digits), exactly as r48c's escalation predicted they would at
their next regen. Aftermath classified to zero unclassified: the 20-file
-tests-closure production family and 17 CRLF phantoms (7 production, 10
-text testdata) restored per the standing rule.

+++ src/core/encoding/binary/binary_test.cs
@@ -45,7 +45,7 @@ internal static ж<Struct> Ꮡs = new(new Struct(
     0x01,
     0x0203,
     0x04050607,
-    (nint)0x08090a0b0c0d0e0fL,
+    0x08090a0b0c0d0e0fL,
     0x10,
     0x1112,
     0x13141516,
